modules: fix prematurely invalidating a packet that has invalid header
We only invalidate the packet if we were in a middle
of a valid one (correct header) and we got a different
header.

The typical case in which this happens is when
the Cell Controller receives the FMPS first
(only fmpsReadLink will accept the packet
and the fofbReadLink will silentyly ignore it),
and then it receives the CC packets (ony fofbReadLink
will accept them and fmpsReadLink will silently ignore
them).